% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolify.r
\name{boolify}
\alias{boolify}
\title{Make boolean variables}
\usage{
boolify(dat, boolvars)
}
\arguments{
\item{dat}{Matrix or data frame containing the data}
}
\description{
Make numeric or character variables boolean
}
