% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_occ_combs.r
\name{co_occ_combs}
\alias{co_occ_combs}
\title{CALCULATE OCCURRENCE COMBINATIONS}
\usage{
co_occ_combs(partition, Y_arr)
}
\arguments{
\item{partition}{A matrix giving the variables from which to form the combinations}

\item{Y}{A list of matrices of (species) occurrences}
}
\description{
Calculate combinations of species occurrences
}
