% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAssociations_modified.r
\name{computeAssociations_modified}
\alias{computeAssociations_modified}
\title{Compute associations (modified)}
\usage{
computeAssociations_modified(hM, start = 1, thin = 1)
}
\arguments{
\item{object}{A fitted Hmsc model object.}

\item{...}{See ?Hmsc::Compute associations for the rest}
}
\description{
Modified function for computing association matrices when using covariate dependent latent variables.
}
