% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePredictedValues_modified.r
\name{computePredictedValues_modified}
\alias{computePredictedValues_modified}
\title{Compute predicted values (modified)}
\usage{
computePredictedValues_modified(
  hM,
  partition = NULL,
  partition.sp = NULL,
  start = 1,
  thin = 1,
  Yc = NULL,
  mcmcStep = 1,
  expected = TRUE,
  initPar = NULL,
  nParallel = 1,
  nChains = length(hM$postList),
  updater = list(),
  verbose = hM$verbose,
  alignPost = TRUE
)
}
\arguments{
\item{hM}{Hmsc object}

\item{partition}{Partition for the cross validation}

\item{alignPost}{boolean}
}
\description{
Modified function for computing predicted values, using the predictHmsc_modified function for predictions
}
