% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories.r
\name{create_directories}
\alias{create_directories}
\title{Create directories}
\usage{
create_directories(dirs, foldname)
}
\arguments{
\item{dirs}{A 'dirlist' object created with set_dirs.}

\item{foldname}{A 'trapfilename' object created with create_name, specifying the name for the study folder.}
}
\description{
Create the folders/directories in which model fits and results will be saved.
}
