% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_name.r
\name{create_name}
\alias{create_name}
\title{Create name}
\usage{
create_name(
  study = "trap17",
  totsamp = NULL,
  nfolds = NULL,
  type = c("fold", "cv", "eval_cv")
)
}
\arguments{
\item{study}{Name of the study}

\item{totsamp}{Total number of MCMC samples}

\item{nfolds}{Number of folds for cross-validation}

\item{obj_type}{Object type: either "fold", "cv" or "eval_cv"}
}
\description{
Create a name for a folder or a cross-validation results file
}
