% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.r
\name{data_prep}
\alias{data_prep}
\title{Data preparation}
\usage{
data_prep(filename = "TRAP17.csv", dirs, save_data = FALSE)
}
\arguments{
\item{filename}{Original data file}

\item{dirs}{A trap17 class "dirlist" object}
}
\description{
Prepare the original data: unify names and define aggregation formulas
}
