% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cv.r
\name{do_cv}
\alias{do_cv}
\title{Perform cross-validation}
\usage{
do_cv(
  ps,
  dat,
  dirs,
  vars,
  higher_eval_levels = TRUE,
  save_cv = TRUE,
  start_iter,
  expect = "both"
)
}
\arguments{
\item{param}{Descr}
}
\description{
Perform cross-validation
}
