% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify.r
\name{dummify}
\alias{dummify}
\title{Make dummy/indicator variables}
\usage{
dummify(dat, dumvars)
}
\arguments{
\item{dat}{Matrix or data frame containing the data}

\item{dumvars}{Character vector containing the names of the variables to be modified}
}
\description{
Transform numeric or character variables with multiple levels into dummy/indicators.
}
