% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_cors.r
\name{higher_cors}
\alias{higher_cors}
\title{Higher-level correlations in predictions
Compute genoptype-population-level correlations in predictions}
\usage{
higher_cors(dat, preds)
}
\arguments{
\item{dat}{Original data}

\item{preds}{Predicted occurrences}
}
\description{
Higher-level correlations in predictions
Compute genoptype-population-level correlations in predictions
}
