% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_colour_palette.r
\name{load_colour_palette}
\alias{load_colour_palette}
\title{Load colour palette}
\usage{
load_colour_palette()
}
\description{
Load a custom, colour-blind-friendly palette of colours for results plotting
}
