% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects_from_dir.r
\name{load_objects_from_dir}
\alias{load_objects_from_dir}
\title{Load objects}
\usage{
load_objects_from_dir(path, study, obj_type)
}
\arguments{
\item{path}{Path to object}

\item{study}{Name of study}

\item{obj_type}{Object type: either "fold", "cv" or "eval_cv"}
}
\description{
Load objects from a directory: model fit, predictions, cross-validation results, ...
}
