% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_and_cv.r
\name{model_and_cv}
\alias{model_and_cv}
\title{Fit models and do cross-validation}
\usage{
model_and_cv(
  dat,
  dirs,
  variants = "ALL",
  sampling,
  returnCVs = FALSE,
  saveCVs = TRUE
)
}
\arguments{
\item{dat}{Descr}

\item{dirs}{Descr}

\item{variants}{Descr}

\item{sampling}{Descr}

\item{returnCVs}{default FALSE}

\item{saveCVs}{default TRUE}
}
\description{
Fit models and do cross-validation
}
