% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.r
\name{pool}
\alias{pool}
\title{Pool trap17 sata}
\usage{
pool(dat, dirs)
}
\arguments{
\item{dirs}{Directories}

\item{save_data}{Save the pooled data}

\item{return_data}{Return the pooled data}
}
\description{
Pool the trap17 data over the two sampling time points
}
