% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictHmsc_modified.r
\name{predictHmsc_modified}
\alias{predictHmsc_modified}
\title{Predict for Hmsc (modified)}
\usage{
predictHmsc_modified(
  object,
  post = poolMcmcChains(object$postList),
  XData = NULL,
  X = NULL,
  XRRRData = NULL,
  XRRR = NULL,
  studyDesign = object$studyDesign,
  ranLevels = object$ranLevels,
  Gradient = NULL,
  Yc = NULL,
  mcmcStep = 1,
  expected = FALSE,
  predictEtaMean = FALSE,
  predictEtaMeanField = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted Hmsc model object.}

\item{post}{A list of posterior samples of the HMSC model. By default uses all samples from the pooled posterior of the hM object.}

\item{...}{See ?Hmsc::predict.Hmsc for the rest}
}
\description{
Modified function for predicting. Fixes the bug when using covariate dependent latent variables.
}
