% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_m.r
\name{prepare_m}
\alias{prepare_m}
\title{Prepare the Hmsc model object}
\usage{
prepare_m(vars, dat)
}
\arguments{
\item{vars}{Variables, created with set_vars}

\item{dat}{Data, i.e. trapdata}
}
\description{
Prepare the Hmsc model object for model fitting
}
