% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.r
\name{process_data}
\alias{process_data}
\title{Process trap17 data}
\usage{
process_data(filename, dirs, save_data = TRUE, return_data = TRUE)
}
\arguments{
\item{filename}{Original data file}

\item{dirs}{A trap17 class "dirlist" object}

\item{save_data}{Save the modified data}

\item{return_data}{Return the pooled data}
}
\description{
Modify variable names and pool the data
}
