% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Hmsc.r
\name{sample_Hmsc}
\alias{sample_Hmsc}
\title{Fit an Hmsc model}
\usage{
sample_Hmsc(dat = dat, vars = vars, dirs = dirs, return_ps = TRUE)
}
\arguments{
\item{param}{Descr}
}
\description{
Fit an Hmsc model with defined variables and save results to defined directories
}
