% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_vars.r
\name{set_vars}
\alias{set_vars}
\title{Set variables for the model}
\usage{
set_vars(study = "trap17", fit, sampling)
}
\arguments{
\item{study}{Name of study}

\item{fit}{Model fit}

\item{sampling}{a "mcmcsettings" object defining the sampling settings}
}
\description{
Load objects from a directory: model fit, predictions, cross-validation results, ...
}
