% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_and_cv.r
\name{model_and_cv}
\alias{model_and_cv}
\title{Fit models and do cross-validation}
\usage{
model_and_cv(
  dat,
  dirs,
  variants = "ALL",
  sampling,
  start_iter,
  do_cv = TRUE,
  saveCVs = TRUE
)
}
\arguments{
\item{dat}{Data}

\item{dirs}{Directories}

\item{variants}{Selection of model variants}

\item{sampling}{Sampling settings}

\item{saveCVs}{default TRUE}
}
\description{
Fit models and optionally does cross-validation. If do_cv = TRUE, returns the performance measures.
}
