% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapdata.r
\docType{data}
\name{trapdata}
\alias{trapdata}
\title{TRAP17 DATA SET}
\format{A list of data frames, character vectors and formulas.
\describe{
  \item{Y}{Virus occurrences.}
  \item{Y_pooled}{Virus occurrences, pooled over the two sampling time points during summer 2017.}
  \item{X}{All independent variables.}
  \item{X_pooled}{All independent variables, pooled over the two sampling time points during summer 2017.}
  \item{yvars}{Names of viruses.}
  \item{xvars}{Names of the independent variables.}
  \item{pool_}{Formulas for pooling the independent variables.}
  \item{poolvar}{The unit over which variables were grouped.}
}}
\usage{
trapdata
}
\description{
The processed trap17 data set of five viruses on 320 sentinel plants in Åland
}
\keyword{datasets}
